requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
telethon==1.34.0
python-dotenv==1.0.0
cryptography==41.0.8
pydantic==2.5.0
asyncio-throttle==1.0.2